Create: api-key auth

+++ server/src/main/java/com/example/server/controller/ContestController.java
@@ -2,6 +2,8 @@ package com.example.server.controller;
 
 import com.example.server.repository.contest.Contest;
 import com.example.server.service.ContestService;
+import com.example.server.service.UserKeyService;
+import jakarta.security.auth.message.AuthException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -10,9 +12,10 @@ import java.util.List;
 @RequestMapping(path = "api/contests")
 public class ContestController {
     private final ContestService contestService;
-
-    public ContestController(ContestService contestService) {
+    private final UserKeyService userKeyService;
+    public ContestController(ContestService contestService, UserKeyService userKeyService) {
         this.contestService = contestService;
+        this.userKeyService = userKeyService;
     }
 
     @GetMapping
@@ -21,17 +24,20 @@ public class ContestController {
     }
 
     @PostMapping
-    public void createContest(@RequestBody Contest contest) {
+    public void createContest(@RequestBody Contest contest, @RequestHeader("x-api-key") String key) throws AuthException { // TODO: x-api-key;
+        userKeyService.checkAccessRights(key, 1);
         contestService.createContest(contest);
     }
 
     @DeleteMapping
-    public void deleteContest(@RequestParam Long id) {
+    public void deleteContest(@RequestParam Long id, @RequestHeader("x-api-key") String key) throws AuthException {
+        userKeyService.checkAccessRights(key, 1);
         contestService.deleteContest(id);
     }
 
     @PutMapping
-    public void updateContest(@RequestBody Contest changes) {
+    public void updateContest(@RequestBody Contest changes, @RequestHeader("x-api-key") String key) throws AuthException {
+        userKeyService.checkAccessRights(key, 1);
         contestService.updateContest(changes);
     }
 }
